% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{clear_old_logs}
\alias{clear_old_logs}
\title{Clear Old Logs}
\usage{
clear_old_logs(keep_months = 6)
}
\arguments{
\item{keep_months}{Number of months of logs to keep.}
}
\value{
Invisible count of files removed.
}
\description{
Removes log files older than a specified number of months.
}
\examples{
\dontrun{
# Keep only last 3 months
clear_old_logs(keep_months = 3)
}
}
