% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{extract_function_calls}
\alias{extract_function_calls}
\title{Extract Function Calls from Expression}
\usage{
extract_function_calls(expr)
}
\arguments{
\item{expr}{An R expression.}
}
\value{
Character vector of function names.
}
\description{
Recursively extracts all function names called in an expression.
}
\keyword{internal}
