% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{generate_hook_content}
\alias{generate_hook_content}
\title{Generate Hook Script Content}
\usage{
generate_hook_content(hook_type)
}
\arguments{
\item{hook_type}{Type of hook}
}
\value{
Character vector of hook script lines
}
\description{
Generate Hook Script Content
}
\keyword{internal}
