% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{generate_precommit_hook}
\alias{generate_precommit_hook}
\title{Generate Pre-commit Hook Script}
\usage{
generate_precommit_hook()
}
\value{
Character vector of script lines
}
\description{
Generate Pre-commit Hook Script
}
\keyword{internal}
