% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_session.R
\name{get_user_input}
\alias{get_user_input}
\title{Get User Input}
\usage{
get_user_input(voice, timeout_secs)
}
\arguments{
\item{voice}{Logical for voice mode}

\item{timeout_secs}{Voice timeout}
}
\value{
Character string of user input
}
\description{
Get User Input
}
\keyword{internal}
