% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{install_single_hook}
\alias{install_single_hook}
\title{Install a Single Git Hook}
\usage{
install_single_hook(hook_dir, hook_type, force)
}
\arguments{
\item{hook_dir}{Path to .git/hooks directory}

\item{hook_type}{Type of hook (e.g., "pre-commit")}

\item{force}{Whether to force overwrite}
}
\value{
Path to installed hook or NULL
}
\description{
Install a Single Git Hook
}
\keyword{internal}
