% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_session.R
\name{is_approval}
\alias{is_approval}
\title{Check for Approval Phrases}
\usage{
is_approval(text)
}
\arguments{
\item{text}{User input text}
}
\value{
Logical
}
\description{
Check for Approval Phrases
}
\keyword{internal}
