% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_session.R
\name{is_exit}
\alias{is_exit}
\title{Check for Exit Phrases}
\usage{
is_exit(text)
}
\arguments{
\item{text}{User input text}
}
\value{
Logical
}
\description{
Detects when user wants to gracefully end the session without
making a decision about the commit.
}
\keyword{internal}
