% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_session.R
\name{is_rejection}
\alias{is_rejection}
\title{Check for Rejection Phrases}
\usage{
is_rejection(text)
}
\arguments{
\item{text}{User input text}
}
\value{
Logical
}
\description{
Check for Rejection Phrases
}
\keyword{internal}
