% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_session.R
\name{output_response}
\alias{output_response}
\title{Output Response}
\usage{
output_response(text, voice)
}
\arguments{
\item{text}{Response text}

\item{voice}{Logical for voice mode}
}
\description{
Output Response
}
\keyword{internal}
