% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_session.R
\name{strip_ansi}
\alias{strip_ansi}
\title{Strip ANSI codes from text}
\usage{
strip_ansi(text)
}
\arguments{
\item{text}{Text with ANSI codes}
}
\value{
Plain text
}
\description{
Strip ANSI codes from text
}
\keyword{internal}
